REFR/C - изменение кодов ошибок

+++ src/common/EdfUtils.c
@@ -18,8 +18,8 @@ size_t strnlength(const char* s, size_t n)
 	return found ? (size_t)(found - s) : n;
 }
 //-----------------------------------------------------------------------------
-size_t minStack = INTPTR_MAX;
-size_t maxStack = INTPTR_MIN;
+size_t minStack = (size_t)INTPTR_MAX;
+size_t maxStack = (size_t)INTPTR_MIN;
 
 int CallStackSize() 
 {
